FileServer registers an io/fs.FS implementation
Errors will propagate through the error handler. For websites
with an HTML frontend, 404's will be rendered as expected.

The io/fs.FS interface is a powerful abstraction. Wrap the fs
with one or more implementations to layer additional functionality.

If the file system also implements the CacheControlFS interface,
a Cache-Control header will be added to the response.

Included is AssetCacheFS which returns an implementation that
applies a uniform "public, max-age=31536000" policy under the
assumption that file names contain a hash of their contents.